Creating inseetion sort (ascending) algorithm. Tried out, and this thing surprisingly works. I am concerned that my AI check in 80% probability, even though I have written the code myself.
Essentially, how insertion sort works, there is a key (aka the very first value we are looking for), which is assumed to be sorted. Then the element of the array that is > than the key is moved ahead and then key value is switched.

I also made sure that array isn't empty, so the sort can carry on.