Writing a menu here....
I still need to write methods for the file selection, as well as the display methods. Though, there is an underlying issue with the fact that there are several file types (256 and 2048), hence I would need to create a counter (boolean data type, most likely) in order to track down the kind of display we need to have.

Maybe I would need to move the whole menu execution into a different file. But it contains the required code for the main execution, so I might not do that.